+++ tests/test_data_bucket.py
@@ -1,5 +1,5 @@
 #!/bin/python3
-import json
+import os
 from jericho.plugin.data_bucket import DataBucket
 
 
@@ -15,19 +15,22 @@ def test_emptying_bucket():
     assert databucket.empty()
     assert databucket.is_empty() == True
     assert databucket.is_full() == False
-    assert databucket.get() == []
 
 
 def test_get_size():
     databucket = DataBucket(100)
     assert databucket.save(("https://google.com", "content")) == True
 
-    # Serialized input ["https://google.com", "content"] in a list = [["https://google.com", "content"]] which is 35
-    assert databucket.get_size() == 35
+    assert databucket.get_size() == 7
 
-
-def test_get_size_is_same_as_get_content_size():
+def test_content_is_ok():
     databucket = DataBucket(100)
     assert databucket.save(("https://google.com", "content")) == True
 
-    assert len(json.dumps(databucket.get())) == databucket.get_size()
+    f = open(f"{databucket.get().replace('.zip', '')}/https:_SLASH__SLASH_google.com", "r")
+    assert f.read() == "content"
+
+def test_zip_file_exist():
+    databucket = DataBucket(100)
+    assert databucket.save(("https://google.com", "content")) == True
+    assert os.path.isfile(databucket.get()) == True